requirements.txt
# requirements.txt
requests==2.31.0
websockets==11.0.3
pyjwt==2.8.0
apscheduler==3.10.4
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.0
pyyaml==6.0.2